requirements.txt

python-dotenv==1.0.1
pandas==2.2.2
numpy==2.4.0
pyyaml==6.0.2
rich==13.7.1
